+++ bikeability/indicators/path_categories.py
@@ -1,12 +1,14 @@
 import logging
 from enum import Enum
-from typing import Dict, Tuple
+from typing import Tuple
 
 import geopandas as gpd
 import pandas as pd
 from shapely import intersects
 from shapely.geometry.multipoint import MultiPoint
 from shapely.ops import split
+
+import bikeability.indicators.path_category_filters as filters
 
 log = logging.getLogger(__name__)
 
@@ -18,6 +20,7 @@ class PathCategory(Enum):
     SHARED_WITH_MOTORISED_TRAFFIC_LOW_SPEED = 'shared_with_cars_up_to_30_km/h'
     SHARED_WITH_MOTORISED_TRAFFIC_MEDIUM_SPEED = 'shared_with_cars_up_to_50_km/h'
     SHARED_WITH_MOTORISED_TRAFFIC_HIGH_SPEED = 'shared_with_cars_above_50_km/h'
+    SHARED_WITH_MOTORISED_TRAFFIC_UNKNOWN_SPEED = 'shared_with_cars_unknown_speed'
     REQUIRES_DISMOUNTING = 'requires_dismounting'
     PEDESTRIAN_EXCLUSIVE = 'pedestrian_exclusive'
     NO_ACCESS = 'no_access'
@@ -40,125 +43,10 @@ class PathCategory(Enum):
         return [category for category in cls if category not in cls.get_not_bikeable()]
 
 
-class PathCategoryFilters:
-    def __init__(self):
-        self.potential_bikeable_highway_values = (
-            'primary',
-            'primary_link',
-            'secondary',
-            'secondary_link',
-            'tertiary',
-            'tertiary_link',
-            'road',
-            'unclassified',
-            'residential',
-            'track',
-            'living_street',
-            'service',
-        )
-
-    def _shared_with_pedestrians(self, d: Dict) -> bool:
-        return (d.get('foot') in ['yes', 'designated'] and d.get('segregated') != 'yes') or (
-            d.get('highway') in ['footway', 'pedestrian', 'path'] and d.get('foot') is None
-        )
-
-    def designated_shared_with_pedestrians(self, d: Dict) -> bool:
-        return (
-            d.get('highway') in ['cycleway', 'path', 'footway', 'pedestrian'] and self._shared_with_pedestrians(d)
-        ) or d.get('highway') == 'track'
-
-    def designated_exclusive(self, d: Dict) -> bool:
-        return d.get('highway') in ['cycleway', 'path', 'footway', 'pedestrian'] and not self._shared_with_pedestrians(
-            d
-        )
-
-    def shared_with_motorised_traffic_walking_speed(self, d: Dict) -> bool:
-        return d.get('highway') in ['living_street', 'service'] or d.get('maxspeed') in [
-            '5',
-            '6',
-            '7',
-            '8',
-            '10',
-            '15',
-            'walk',
-        ]
-
-    def shared_with_motorised_traffic_low_speed(self, d: Dict) -> bool:
-        return d.get('maxspeed') in ['20', '25', '30'] or d.get('zone:maxspeed') in ['DE:30', '30']
-
-    def shared_with_motorised_traffic_medium_speed(self, d: Dict) -> bool:
-        return (
-            d.get('maxspeed') in ['35', '40', '45', '50', 'DE:urban', 'AT:urban']
-            or d.get('maxspeed:forward') in ['35', '40', '45', '50', 'DE:urban', 'AT:urban']
-            or d.get('maxspeed:type') in ['DE:urban', 'AT:urban']
-            or d.get('zone:maxspeed') in ['DE:urban', 'AT:urban']
-            or d.get('highway') == 'residential'
-        )
-
-    def shared_with_motorised_traffic_high_speed(self, d: Dict) -> bool:
-        return (
-            d.get('maxspeed') in ['60', '70', '80', '90', '100', 'DE:rural', 'AT:rural']
-            or d.get('maxspeed:forward') in ['60', '70', '80', '90', '100', 'DE:rural', 'AT:rural']
-            or d.get('maxspeed:type') in ['DE:rural', 'AT:rural']
-            or d.get('zone:maxspeed') in ['DE:rural', 'AT:rural']
-        ) or d.get('highway') == 'unclassified'
-
-    def requires_dismounting(self, d: Dict) -> bool:
-        return (
-            d.get('bicycle') == 'dismount'
-            or '1012-32' in d.get('traffic_sign', 'no')
-            or (
-                d.get('highway') == 'steps'
-                and (
-                    d.get('ramp:bicycle') == 'yes' or d.get('ramp') == 'yes' or d.get('ramp:wheelchair') == 'yes',
-                    d.get('ramp:stroller') == 'yes',
-                )
-            )
-            or d.get('railway') == 'platform'
-            or d.get('highway') == 'platform'
-            or d.get('ford') is not None
-        )
-
-    def pedestrian_exclusive(self, d: Dict) -> bool:
-        return (
-            d.get('highway') in ['footway', 'pedestrian']
-            and (
-                d.get('bicycle')
-                not in [
-                    'yes',
-                    'designated',
-                    'dismount',
-                ]
-                or d.get('bicycle:conditional') is not None
-            )
-            and (d.get('access') not in ['no', 'private', 'permit', 'military', 'delivery', 'customers', 'emergency'])
-        )
-
-    ###############or d.get('highway') == 'footway'
-    ##############and (d.get('bicycle') not in ['yes', 'designated', 'permissive'])        #### this one shoul dbe added?
-
-    def restricted_access(self, d: Dict) -> bool:
-        return (
-            d.get('access') in ['no', 'private', 'permit', 'military', 'delivery', 'customers', 'emergency']
-            or d.get('motorroad') == 'yes'
-            or d.get('bicycle') in ['no', 'private', 'use_sidepath', 'discouraged', 'destination']
-            or d.get('highway')
-            not in [
-                *self.potential_bikeable_highway_values,
-                'pedestrian',
-                'path',
-                'cycleway',
-                'footway',
-                'steps',
-                'platform',
-            ]
-        )
-
-
 def apply_path_category_filters(row: pd.Series) -> PathCategory:
-    filters = PathCategoryFilters()
-
-    match row['@other_tags']:
+    tags = row['@other_tags']
+    speed_limit = filters.parse_maxspeed_tag(tags)
+    match tags:
         case x if filters.requires_dismounting(x):
             return PathCategory.REQUIRES_DISMOUNTING
         case x if filters.pedestrian_exclusive(x):
@@ -169,14 +57,16 @@ def apply_path_category_filters(row: pd.Series) -> PathCategory:
             return PathCategory.EXCLUSIVE
         case x if filters.designated_shared_with_pedestrians(x):
             return PathCategory.SHARED_WITH_PEDESTRIANS
-        case x if filters.shared_with_motorised_traffic_walking_speed(x):
+        case x if filters.shared_with_motorised_traffic_walking_speed(x, speed_limit):
             return PathCategory.SHARED_WITH_MOTORISED_TRAFFIC_WALKING_SPEED
-        case x if filters.shared_with_motorised_traffic_low_speed(x):
+        case x if filters.shared_with_motorised_traffic_low_speed(x, speed_limit):
             return PathCategory.SHARED_WITH_MOTORISED_TRAFFIC_LOW_SPEED
-        case x if filters.shared_with_motorised_traffic_medium_speed(x):
+        case x if filters.shared_with_motorised_traffic_medium_speed(x, speed_limit):
             return PathCategory.SHARED_WITH_MOTORISED_TRAFFIC_MEDIUM_SPEED
-        case x if filters.shared_with_motorised_traffic_high_speed(x):
+        case x if filters.shared_with_motorised_traffic_high_speed(x, speed_limit):
             return PathCategory.SHARED_WITH_MOTORISED_TRAFFIC_HIGH_SPEED
+        case x if filters.shared_with_motorised_traffic_unknown_speed(x, speed_limit):
+            return PathCategory.SHARED_WITH_MOTORISED_TRAFFIC_UNKNOWN_SPEED
         case _:
             return PathCategory.UNKNOWN